Move to container-aware JSON objects and arrays
If we want to handle cases where multiple JSON models are used at the
same time, we need to know which JSON object/array is "owned" by which
model. JSON libraries typically do not provide navigable child->parent
relationships, so we have to add our own.

This is done by replacing JSONObject and JSONArray with JsonModelObject
and JsonModelArray, which are proxies of a JSONObject and JSONArray
which intercept all calls in order to keep track of the last object that
contained them.

A JSON object contains objects that are values of one of their keys, like
{"y": 1} in this case:

  {"x": {"y": 1}}

A JSON array may contain JSON objects within its values:

  [{"y": 1}, 2, 3]

A JsonModel acts as the direct container of its root JSON object or
array.

Under these rules, it is then easy to check which model last "owned" a
given JSON object/array, by repeatedly moving up through the
child->parent relationship until we reach the model.