Allow multiple containers in JSON objects/arrays
JSON does not impose the constraint of always having a single container,
like EMF. It is possible to reuse the same object from multiple places
in a document while we are building it, which means it would have
multiple containers.

This means that in principle, we could have cases where a JSON object
could live across multiple models.

This commit changes the Contained interface to allow for a set of
containers rather than just one, and adds tests checking for this sort
of object reuse. When this object reuse is undesirable, users can call
the JsonModel.deepClone(x) method to obtain a deep clone which is
independent from its original copy.